Explicit conversion of length back to int to avoid 64 bit warning.

+++ lib/core/libwebsockets.c
@@ -149,7 +149,7 @@ lws_hex_to_byte_array(const char *h, uint8_t *dest, int max)
 	if (max < 0)
 		return -1;
 
-	return dest - odest;
+	return (int)(dest - odest);
 }
